+++ b2/console_tool.py
@@ -74,6 +74,8 @@ from b2sdk.v2 import (
     parse_sync_folder,
     ReplicationMonitor,
     ProgressReport,
+    TwoWayReplicationCheckGenerator,
+    CheckState,
 )
 from b2sdk.v2.exception import (
     B2Error,
@@ -188,6 +190,25 @@ def apply_or_none(fcn, value):
         return None
     else:
         return fcn(value)
+
+
+def to_human_readable(value: Any) -> str:
+    if isinstance(value, Enum):
+        return value.name
+
+    if isinstance(value, bool):
+        return 'Yes' if value else 'No'
+
+    if isinstance(value, Bucket):
+        return value.name
+
+    if isinstance(value, ApplicationKey):
+        return value.id_
+
+    if value is None:
+        return ''
+
+    return str(value)
 
 
 class DescriptionGetter:
@@ -2569,10 +2590,93 @@ class ReplicationUnpause(ReplicationRuleChanger):
 
 
 @B2.register_subcommand
+class ReplicationInspect(Command):
+    """
+    Detect possible misconfigurations of replication by analyzing
+    replication rules, buckets and keys.
+
+    --output-format
+    "Console" output format is meant to be human-readable and is subject to change
+    in any further release. One should use "json" for reliable "no-breaking-changes"
+    output format.
+    """
+
+    @classmethod
+    def _setup_parser(cls, parser):
+        super()._setup_parser(parser)
+        parser.add_argument('--source-profile', metavar='SOURCE_PROFILE')
+        parser.add_argument('--destination-profile', metavar='DESTINATION_PROFILE')
+        parser.add_argument('--source-bucket', metavar='SOURCE_BUCKET_NAME')
+        parser.add_argument('--destination-bucket', metavar='DESTINATION_BUCKET_NAME')
+        parser.add_argument('--rule', metavar='REPLICATION_RULE_NAME')
+        parser.add_argument('--file-name-prefix', metavar='FILE_NAME_PREFIX')
+        parser.add_argument('--show-all-checks', action='store_true')
+
+        parser.add_argument('--output-format', default='console', choices=('console', 'json'))
+
+    def run(self, args):
+        source_api = _get_b2api_for_profile(args.source_profile)
+        destination_api = _get_b2api_for_profile(args.destination_profile or args.source_profile)
+
+        troubleshooter = TwoWayReplicationCheckGenerator(
+            source_api=source_api,
+            destination_api=destination_api,
+            filter_source_bucket_name=args.source_bucket,
+            filter_destination_bucket_name=args.destination_bucket,
+            filter_replication_rule_name=args.rule,
+            file_name_prefix=args.file_name_prefix,
+        )
+
+        results = [check.as_dict() for check in troubleshooter.iter_checks()]
+
+        if args.output_format == 'json':
+            self._print_json(
+                [
+                    {key: to_human_readable(value)
+                     for key, value in result.items()} for result in results
+                ]
+            )
+        elif args.output_format == 'console':
+            self._print_console(results, show_all_checks=args.show_all_checks)
+        else:
+            self._print_stderr(f'ERROR: format "{args.output_format}" is not supported')
+            return 1
+
+        return 0
+
+    def _print_console(self, results: List[dict], show_all_checks: bool = False) -> None:
+        for result in results:
+
+            # print keys starting with `_` as text before table
+            self._print('Configuration:')
+            for key, value in result.items():
+                if key.startswith('_'):
+                    self._print(
+                        ' ' * 2 + key[1:].replace('_', ' ') + ': ' + to_human_readable(value)
+                    )
+
+            # print other keys as rows rows
+            rows = {
+                key.replace('_', ' '): to_human_readable(value)
+                for key, value in result.items()
+                if not key.startswith('_') and (value != CheckState.OK or show_all_checks)
+            }.items()
+            self._print('Checks:')
+
+            key = None
+            for key, value in rows:
+                self._print(' ' * 2 + key + ': ' + value)
+            if not key:  # loop was not entered
+                self._print(' ' * 2 + '-')
+
+            self._print('')
+
+
+@B2.register_subcommand
 class ReplicationStatus(Command):
     """
-    Inspects files in only source or both source and destination buckets
-    (potentially from different accounts) and provides detailed replication statistics.
+    Inspect files in only source or both source and destination buckets
+    (potentially from different accounts) and provide detailed replication statistics.
 
     Please be aware that only latest file versions are inspected, so any previous
     file versions are not represented in these statistics.
@@ -2649,13 +2753,14 @@ class ReplicationStatus(Command):
             }
 
         if args.output_format == 'json':
-            self.output_json(results)
+            self._print_json(results)
         elif args.output_format == 'console':
-            self.output_console(results)
+            self._print_console(results)
         elif args.output_format == 'csv':
             self.output_csv(results)
         else:
             self._print_stderr(f'ERROR: format "{args.output_format}" is not supported')
+            return 1
 
         return 0
 
@@ -2683,29 +2788,13 @@ class ReplicationStatus(Command):
     def filter_results_columns(cls, results: List[dict], columns: List[str]) -> List[dict]:
         return [{key: result[key] for key in columns} for result in results]
 
-    @classmethod
-    def to_human_readable(cls, value: Any) -> str:
-        if isinstance(value, Enum):
-            return value.name
-
-        if isinstance(value, bool):
-            return 'Yes' if value else 'No'
-
-        if value is None:
-            return ''
-
-        return str(value)
-
-    def output_json(self, results: Dict[str, List[dict]]) -> None:
-        self._print_json(results)
-
-    def output_console(self, results: Dict[str, List[dict]]) -> None:
+    def _print_console(self, results: Dict[str, List[dict]]) -> None:
         for rule_name, rule_results in results.items():
             self._print(f'Replication "{rule_name}":')
             rule_results = [
                 {
                     key.replace('_', '\n'):  # split key to minimize column size
-                    self.to_human_readable(value)
+                    to_human_readable(value)
                     for key, value in result.items()
                 } for result in rule_results
             ]
@@ -2721,7 +2810,7 @@ class ReplicationStatus(Command):
                     'rule name': rule_name,
                     **{
                         key.replace('_', '\n'):  # split key to minimize column size
-                        self.to_human_readable(value)
+                        to_human_readable(value)
                         for key, value in result.items()
                     },
                 } for result in rule_results